+++ CHANGELOG.md
@@ -182,3 +182,4 @@
 - 2023-11-23: Shutdown ready pool, qualify lead, lead, users, +5
 - 2023-11-24: Init conf, bettercontact, follow up, qualify lead, +3
 - 2023-11-24: Docs configuration, conf, contacts, leads, +2
+- 2023-11-25: Shutdown mod, discovery, deal, discover, +5